refactor finalize_block token updates
