+++ crates/token/Cargo.toml
@@ -19,5 +19,5 @@ multicore = ["namada_shielded_token/multicore"]
 [dependencies]
 namada_core = { path = "../core" }
 namada_shielded_token = { path = "../shielded_token" }
-namada_storage = { path = "../storage" }
+namada_state = { path = "../state" }
 namada_trans_token = { path = "../trans_token" }
